Update 2.6
Changed the title on the bottom of the page to be a little more serious.
Added the underwater settings menu project in the apps section on the studies page. (also two more projects made in figma have been added but are currently hidden.)
Removed "Ervaring" page.

Bug fixes:
- Eased the severity of the back-to-top button jumping on the document pages.

I might improve upon:
- Skills kan een klein tekstje gebruiken
- Eventueel meer tekeningen toevoegen op de hobbies pagina (Piechart?)
- Contact pagina kan nog een kaart gebruiken
- Ik kan de andere apps ook toevoegen maar dit kan mogelijk mijn vaardigheden vals vertonen.